Add internal cyclization groups on a phenyl ring
R3_B and R4_B added representing internal cyclizations
on a phenyl ring, which should be extremely slow
but is necessary to discover reverse direction kinetics

Banned the R5_B long route transition state which is
a functional equivalent of R3.

Ban the phenyl isomer internal cyclizations which
are functionally identical to R3_B and R4_B cyclizations.